ajust if table exist knex migration

+++ src/database/knex/migrations/20240311190703_createTags.js
@@ -1,9 +1,14 @@
-exports.up = knex => knex.schema.createTableIfNotExists("tags", table => {
-  table.increments("id")
-  table.text("name").notNullable()
+exports.up = async knex =>{
+  const exist = await knex.schema.hasTable("tags")
 
-  table.integer("note_id").references("id").inTable("notes").onDelete("CASCADE")
-  table.integer("user_id").references("id").inTable("users")
-});
-
-exports.down = knex => knex.schema.dropTableIfExists("tags");
+  if(!exist){
+    knex.schema.createTable("tags", table => {
+      table.increments("id")
+      table.text("name").notNullable()
+      table.integer("note_id").references("id").inTable("notes").onDelete("CASCADE")
+      table.integer("user_id").references("id").inTable("users")
+    }) 
+  }
+};
+  
+exports.down = async knex => await knex.schema.dropTableIfExists("tags");
