+++ src/database/knex/migrations/20240311191154_createLinks.js
@@ -1,9 +1,14 @@
-exports.up = knex => knex.schema.createTableIfNotExists("links", table => {
-  table.increments("id")
-  table.text("url").notNullable()
+exports.up = async knex => {
+  const exist = await knex.schema.hasTable("links")
 
-  table.integer("note_id").references("id").inTable("notes").onDelete("CASCADE")
-  table.timestamp("created_at").default(knex.fn.now())
-});
+  if(!exist){
+    await knex.schema.createTable("links", table => {
+      table.increments("id")
+      table.text("url").notNullable()
+      table.integer("note_id").references("id").inTable("notes").onDelete("CASCADE")
+      table.timestamp("created_at").default(knex.fn.now())
+  })
+} 
+};
 
-exports.down = knex => knex.schema.dropTableIfExists("links");
+exports.down = async knex => await knex.schema.dropTableIfExists("links");
